Adds a limit for when conformers crash from the first chunk
Because conformers are built from experimental chunks, it is possible
that the first chunk, especially if it is large, does not pass our
validation routines. A maximum number of trials was added to the construction
loop - set to 5.

After this maximum, the process aborts production. It can happen that
a given number of conformers built successfully before an eventual break.

5 was choosen for no particular reason, I needed to give a number. This can
change in the future if we see an higher number fits better.